Homing missiles aim based on nearest target to missile, not by closest target to generator
